// src/lib/types/ui.ts
/** Identifier of an UI element */
export enum UI {
	// CPU
	PROGRAM_COUNTER = "PROGRAM_COUNTER",
	PROGRAM_COUNTER_INCREMENT = "PROGRAM_COUNTER_INCREMENT",
	INSTRUCTION_REGISTER = "INSTRUCTION_REGISTER",
	INSTRUCTION_REGISTER_OPERAND = "INSTRUCTION_REGISTER_OPERAND",
	INSTRUCTION_REGISTER_OPCODE = "INSTRUCTION_REGISTER_OPCODE",
	DECODER = "DECODER",
	CONTROL_UNIT = "CONTROL_UNIT",
	MUX = "MUX",
	ALU_OPERAND_1 = "ALU_OPERAND_1",
	ALU_OPERAND_2 = "ALU_OPERAND_2",
	ALU_OPERATION = "ALU_OPERATION",
	ACCUMULATOR = "ACCUMULATOR",
	STATUS_WORD = "STATUS_WORD",
	STATUS_WORD_ZERO_FLAG = "STATUS_WORD_ZERO_FLAG",
	STATUS_WORD_NEGATIVE_FLAG = "STATUS_WORD_NEGATIVE_FLAG",
	// Memory
	MEMORY_SELECTED_ADDRESS = "MEMORY_SELECTED_ADDRESS",
	MEMORY_SELECTED_DATA = "MEMORY_SELECTED_DATA",
	// Labels
	LABEL = "LABEL",
	// Control bar
	CONTROL_BAR = "CONTROL_BAR",
	// Program execution
	PROGRAM_EXECUTION_BUTTON_GROUP = "PROGRAM_EXECUTION_BUTTON_GROUP",
	RESET_EXECUTION_BUTTON = "RESET_EXECUTION_BUTTON",
	TOGGLE_PROGRAM_EXECUTION_BUTTON = "TOGGLE_PROGRAM_EXECUTION_BUTTON",
	FAST_PROGRAM_EXECUTION_BUTTON = "FAST_PROGRAM_EXECUTION_BUTTON",
	// Instruction execution
	INSTRUCTION_EXECUTION_BUTTON_GROUP = "INSTRUCTION_EXECUTION_BUTTON_GROUP",
	TOGGLE_INSTRUCTION_EXECUTION_BUTTON = "TOGGLE_INSTRUCTION_EXECUTION_BUTTON",
	FAST_INSTRUCTION_EXECUTION_BUTTON = "FAST_INSTRUCTION_EXECUTION_BUTTON",
	// Step execution
	STEP_EXECUTION_BUTTON_GROUP = "STEP_EXECUTION_BUTTON_GROUP",
	TOGGLE_STEP_EXECUTION_BUTTON = "TOGGLE_STEP_EXECUTION_BUTTON",
	FAST_STEP_EXECUTION_BUTTON = "FAST_STEP_EXECUTION_BUTTON",
	// Execution speed
	EXECUTION_SPEED_SLIDER = "EXECUTION_SPEED_SLIDER",
	// Binary mode toggle
	BINARY_MODE_TOGGLE = "BINARY_MODE_TOGGLE",
	// Animations toggle
	ANIMATIONS_TOGGLE = "ANIMATIONS_TOGGLE",
}